added basic navigation & buoy/gate capabilities that were copied across files to task server base, works
