+++ all_seaing_autonomy/all_seaing_autonomy/roboboat/delivery_qual.py
@@ -35,9 +35,6 @@ class DeliveryQual(TaskServerBase):
         )
 
         bringup_prefix = get_package_share_directory("all_seaing_bringup")
-        self.declare_parameter("is_sim", False)
-
-        self.is_sim = self.get_parameter("is_sim").get_parameter_value().bool_value
 
         # update from subs
         self.height = None
