+++ all_seaing_autonomy/all_seaing_autonomy/roboboat/docking_fallback.py
@@ -88,7 +88,6 @@ class DockingFallback(TaskServerBase):
         self.time_last_seen_buoys = self.get_clock().now().nanoseconds / 1e9
 
         bringup_prefix = get_package_share_directory("all_seaing_bringup")
-        self.declare_parameter("is_sim", False)
         
         # update from subs
         self.dock_banners = []
@@ -103,8 +102,6 @@ class DockingFallback(TaskServerBase):
 
         self.state = None
         self.result = False
-
-        self.is_sim = self.get_parameter("is_sim").get_parameter_value().bool_value
 
         self.declare_parameter(
             "shape_label_mappings_file",
