+++ all_seaing_autonomy/all_seaing_autonomy/roboboat/follow_buoy_pid.py
@@ -97,9 +97,6 @@ class FollowBuoyPID(TaskServerBase):
 
 
         bringup_prefix = get_package_share_directory("all_seaing_bringup")
-        self.declare_parameter("is_sim", False)
-
-        self.is_sim = self.get_parameter("is_sim").get_parameter_value().bool_value
 
         # update from subs
         self.height = None
