+++ all_seaing_autonomy/all_seaing_autonomy/roboboat/harbor_alert.py
@@ -37,14 +37,11 @@ class HarborAlert(TaskServerBase):
         self.declare_parameter("stationkeep_dist", 2.0)
         self.stationkeep_dist = self.get_parameter("stationkeep_dist").get_parameter_value().double_value
 
-        self.declare_parameter("probe_distance", 10)
         self.declare_parameter("adaptive_distance", 0.7)
         self.adaptive_distance = self.get_parameter("adaptive_distance").get_parameter_value().double_value
 
         self.declare_parameter("duplicate_dist", 0.5)
         self.duplicate_dist = self.get_parameter("duplicate_dist").get_parameter_value().double_value
-
-        self.declare_parameter("is_sim", False)
 
         self.blue_buoy_pos = np.array([0.0, 0.0])
         self.return_pos = None
@@ -84,7 +81,7 @@ class HarborAlert(TaskServerBase):
     def should_accept_task(self, goal_request=None):
         if self.obstacles is None:
             return False
-        return self.blue_buoy_detected()
+        return self.buoy_detected(self.obstacles, self.blue_labels, self.duplicate_dist)
     
     def goal_callback(self, goal_request):
         self.get_logger().info(f'Task Server [{self.server_name}] received task')
@@ -97,7 +94,7 @@ class HarborAlert(TaskServerBase):
         return GoalResponse.ACCEPT
 
     def init_setup(self):
-        if self.blue_buoy_detected():
+        if self.buoy_detected(self.obstacles, self.blue_labels, self.duplicate_dist):
             self.state = ReturnState.STATIONKEEPING
         else:
             self.state = ReturnState.RETURNING
@@ -126,7 +123,7 @@ class HarborAlert(TaskServerBase):
         buoy_direction = robot_buoy_vector / robot_buoy_dist
 
         def update_current_point():
-            self.blue_buoy_detected()
+            self.buoy_detected(self.obstacles, self.blue_labels, self.duplicate_dist)
             return self.blue_buoy_pos - self.stationkeep_dist * buoy_direction
 
         self.stationkeep_point = update_current_point()
@@ -149,34 +146,6 @@ class HarborAlert(TaskServerBase):
         # self.move_to_waypoint((self.return_pos[0], self.return_pos[1], math.atan2(self.return_dir[1], self.return_dir[0])), busy_wait=True)
         
         return Task.Result(success=True)
-
-    def blue_buoy_detected(self, buoy_front=False):
-        '''
-        Check if the blue buoy for turning is detected (returns boolean).
-        Also sets the position of the blue buoy if it is found.
-        '''
-        # backup_buoy = None
-        updated_pos = False
-        for obstacle in self.obstacles:
-            if obstacle.label in self.blue_labels:
-                ob_pos = ob_coords(obstacle)
-                if np.linalg.norm(self.robot_pos - ob_pos) > self.buoy_dist_thres:
-                    continue
-                buoy_dir = ob_pos - self.robot_pos
-                dot_prod = buoy_dir @ self.robot_dir
-                # if (backup_buoy is None) or (self.buoy_found and (self.norm(self.blue_buoy_pos, buoy_pos) < self.norm(self.blue_buoy_pos, backup_buoy))):
-                #     backup_buoy = buoy_pos
-                if ((not buoy_front) or (dot_prod > 0)) and ((not self.buoy_found) or (np.linalg.norm(self.blue_buoy_pos - ob_pos) < self.duplicate_dist)):
-                    self.buoy_found = True
-                    updated_pos = True
-                    self.blue_buoy_pos = ob_pos
-                    break
-        # if (not updated_pos) and (backup_buoy is not None):
-        #     self.get_logger().info('SWITCHING TO BACKUP BUOY')
-        #     self.blue_buoy_pos = backup_buoy
-        return self.buoy_found
-
-
 
 def main(args=None):
     rclpy.init(args=args)
